# api-server/generated/api/company/api.gen.go
//go:build go1.22

// Package company provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package company

import (
	"bytes"
	"context"
	"encoding/json"
	"fmt"
	"net/http"
	"time"

	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Defines values for TsunaguWorksProjectStatus.
const (
	Closed    TsunaguWorksProjectStatus = "closed"
	Draft     TsunaguWorksProjectStatus = "draft"
	Published TsunaguWorksProjectStatus = "published"
)

// Valid indicates whether the value is a known member of the TsunaguWorksProjectStatus enum.
func (e TsunaguWorksProjectStatus) Valid() bool {
	switch e {
	case Closed:
		return true
	case Draft:
		return true
	case Published:
		return true
	default:
		return false
	}
}

// TsunaguWorksApiError エラーレスポンスの共通形
type TsunaguWorksApiError struct {
	Error     string  `json:"error"`
	RequestId *string `json:"request_id,omitempty"`
}

// TsunaguWorksAuthToken 認証トークン（web は httpOnly Cookie に変換して保持し、ブラウザ JS には渡さない）
type TsunaguWorksAuthToken struct {
	Token string `json:"token"`
}

// TsunaguWorksCompanyMe ログイン中の企業ユーザー情報
type TsunaguWorksCompanyMe struct {
	Description string `json:"description"`
	Email       string `json:"email"`
	Location    string `json:"location"`
	Name        string `json:"name"`
	UserId      int64  `json:"user_id"`
}

// TsunaguWorksCompanySignupInput 企業のサインアップ入力（アカウントとプロフィールを同時に作る）
type TsunaguWorksCompanySignupInput struct {
	Description *string             `json:"description,omitempty"`
	Email       openapi_types.Email `json:"email"`
	Location    *string             `json:"location,omitempty"`

	// Name 会社名
	Name string `json:"name"`

	// Password パスワード（上限は bcrypt の入力上限 72 バイトに合わせる）
	Password string `json:"password"`
}

// TsunaguWorksLoginInput ログイン入力
type TsunaguWorksLoginInput struct {
	Email    openapi_types.Email `json:"email"`
	Password string              `json:"password"`
}

// TsunaguWorksProject 案件（企業が掲載する仕事）
type TsunaguWorksProject struct {
	CreatedAt   time.Time `json:"created_at"`
	Description string    `json:"description"`

	// HourlyRateMax 想定時給の上限（円）。未設定は null
	HourlyRateMax *int32 `json:"hourly_rate_max"`

	// HourlyRateMin 想定時給の下限（円）。未設定は null
	HourlyRateMin *int32 `json:"hourly_rate_min"`

	// HoursPerWeek 週の想定稼働時間
	HoursPerWeek int32 `json:"hours_per_week"`
	Id           int64 `json:"id"`
	RemoteOk     bool  `json:"remote_ok"`

	// RequiredSkills 必須スキル（例: ["Go", "PostgreSQL"]）
	RequiredSkills []string `json:"required_skills"`

	// Status 案件の掲載状態
	Status TsunaguWorksProjectStatus `json:"status"`
	Title  string                    `json:"title"`
}

// TsunaguWorksProjectCreateInput 案件の作成入力（status はサーバー側で draft 固定のため含めない）
type TsunaguWorksProjectCreateInput struct {
	Description    string   `json:"description"`
	HourlyRateMax  *int32   `json:"hourly_rate_max,omitempty"`
	HourlyRateMin  *int32   `json:"hourly_rate_min,omitempty"`
	HoursPerWeek   int32    `json:"hours_per_week"`
	RemoteOk       bool     `json:"remote_ok"`
	RequiredSkills []string `json:"required_skills"`
	Title          string   `json:"title"`
}

// TsunaguWorksProjectStatus 案件の掲載状態
type TsunaguWorksProjectStatus string

// AuthLoginJSONRequestBody defines body for AuthLogin for application/json ContentType.
type AuthLoginJSONRequestBody = TsunaguWorksLoginInput

// AuthSignupJSONRequestBody defines body for AuthSignup for application/json ContentType.
type AuthSignupJSONRequestBody = TsunaguWorksCompanySignupInput

// ProjectsCreateJSONRequestBody defines body for ProjectsCreate for application/json ContentType.
type ProjectsCreateJSONRequestBody = TsunaguWorksProjectCreateInput

// ServerInterface represents all server handlers.
type ServerInterface interface {

	// (POST /auth/login)
	AuthLogin(w http.ResponseWriter, r *http.Request)

	// (GET /auth/me)
	AuthMe(w http.ResponseWriter, r *http.Request)

	// (POST /auth/signup)
	AuthSignup(w http.ResponseWriter, r *http.Request)

	// (POST /projects)
	ProjectsCreate(w http.ResponseWriter, r *http.Request)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// AuthLogin operation middleware
func (siw *ServerInterfaceWrapper) AuthLogin(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.AuthLogin(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// AuthMe operation middleware
func (siw *ServerInterfaceWrapper) AuthMe(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.AuthMe(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// AuthSignup operation middleware
func (siw *ServerInterfaceWrapper) AuthSignup(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.AuthSignup(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ProjectsCreate operation middleware
func (siw *ServerInterfaceWrapper) ProjectsCreate(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ProjectsCreate(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{})
}

// ServeMux is an abstraction of [http.ServeMux].
type ServeMux interface {
	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
	http.Handler
}

type StdHTTPServerOptions struct {
	BaseURL          string
	BaseRouter       ServeMux
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, m ServeMux) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseRouter: m,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, m ServeMux, baseURL string) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseURL:    baseURL,
		BaseRouter: m,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options StdHTTPServerOptions) http.Handler {
	m := options.BaseRouter

	if m == nil {
		m = http.NewServeMux()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	m.HandleFunc(http.MethodPost+" "+options.BaseURL+"/auth/login", wrapper.AuthLogin)
	m.HandleFunc(http.MethodGet+" "+options.BaseURL+"/auth/me", wrapper.AuthMe)
	m.HandleFunc(http.MethodPost+" "+options.BaseURL+"/auth/signup", wrapper.AuthSignup)
	m.HandleFunc(http.MethodPost+" "+options.BaseURL+"/projects", wrapper.ProjectsCreate)

	return m
}

type AuthLoginRequestObject struct {
	Body *AuthLoginJSONRequestBody
}

type AuthLoginResponseObject interface {
	VisitAuthLoginResponse(w http.ResponseWriter) error
}

type AuthLogin200JSONResponse TsunaguWorksAuthToken

func (response AuthLogin200JSONResponse) VisitAuthLoginResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type AuthLogindefaultJSONResponse struct {
	Body       TsunaguWorksApiError
	StatusCode int
}

func (response AuthLogindefaultJSONResponse) VisitAuthLoginResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response.Body); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(response.StatusCode)
	_, err := buf.WriteTo(w)
	return err
}

type AuthMeRequestObject struct {
}

type AuthMeResponseObject interface {
	VisitAuthMeResponse(w http.ResponseWriter) error
}

type AuthMe200JSONResponse TsunaguWorksCompanyMe

func (response AuthMe200JSONResponse) VisitAuthMeResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type AuthMedefaultJSONResponse struct {
	Body       TsunaguWorksApiError
	StatusCode int
}

func (response AuthMedefaultJSONResponse) VisitAuthMeResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response.Body); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(response.StatusCode)
	_, err := buf.WriteTo(w)
	return err
}

type AuthSignupRequestObject struct {
	Body *AuthSignupJSONRequestBody
}

type AuthSignupResponseObject interface {
	VisitAuthSignupResponse(w http.ResponseWriter) error
}

type AuthSignup201JSONResponse TsunaguWorksAuthToken

func (response AuthSignup201JSONResponse) VisitAuthSignupResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(201)
	_, err := buf.WriteTo(w)
	return err
}

type AuthSignupdefaultJSONResponse struct {
	Body       TsunaguWorksApiError
	StatusCode int
}

func (response AuthSignupdefaultJSONResponse) VisitAuthSignupResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response.Body); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(response.StatusCode)
	_, err := buf.WriteTo(w)
	return err
}

type ProjectsCreateRequestObject struct {
	Body *ProjectsCreateJSONRequestBody
}

type ProjectsCreateResponseObject interface {
	VisitProjectsCreateResponse(w http.ResponseWriter) error
}

type ProjectsCreate201JSONResponse TsunaguWorksProject

func (response ProjectsCreate201JSONResponse) VisitProjectsCreateResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(201)
	_, err := buf.WriteTo(w)
	return err
}

type ProjectsCreatedefaultJSONResponse struct {
	Body       TsunaguWorksApiError
	StatusCode int
}

func (response ProjectsCreatedefaultJSONResponse) VisitProjectsCreateResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response.Body); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(response.StatusCode)
	_, err := buf.WriteTo(w)
	return err
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {

	// (POST /auth/login)
	AuthLogin(ctx context.Context, request AuthLoginRequestObject) (AuthLoginResponseObject, error)

	// (GET /auth/me)
	AuthMe(ctx context.Context, request AuthMeRequestObject) (AuthMeResponseObject, error)

	// (POST /auth/signup)
	AuthSignup(ctx context.Context, request AuthSignupRequestObject) (AuthSignupResponseObject, error)

	// (POST /projects)
	ProjectsCreate(ctx context.Context, request ProjectsCreateRequestObject) (ProjectsCreateResponseObject, error)
}

type StrictHandlerFunc func(ctx context.Context, w http.ResponseWriter, r *http.Request, request any) (any, error)
type StrictMiddlewareFunc func(f StrictHandlerFunc, operationID string) StrictHandlerFunc

type StrictHTTPServerOptions struct {
	RequestErrorHandlerFunc  func(w http.ResponseWriter, r *http.Request, err error)
	ResponseErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: StrictHTTPServerOptions{
		RequestErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		},
		ResponseErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		},
	}}
}

func NewStrictHandlerWithOptions(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc, options StrictHTTPServerOptions) ServerInterface {
	if options.RequestErrorHandlerFunc == nil {
		options.RequestErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	if options.ResponseErrorHandlerFunc == nil {
		options.ResponseErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		}
	}
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: options}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
	options     StrictHTTPServerOptions
}

// AuthLogin operation middleware
func (sh *strictHandler) AuthLogin(w http.ResponseWriter, r *http.Request) {
	var request AuthLoginRequestObject

	var body AuthLoginJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
		return
	}
	request.Body = &body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.AuthLogin(ctx, request.(AuthLoginRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "AuthLogin")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(AuthLoginResponseObject); ok {
		if err := validResponse.VisitAuthLoginResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// AuthMe operation middleware
func (sh *strictHandler) AuthMe(w http.ResponseWriter, r *http.Request) {
	var request AuthMeRequestObject

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.AuthMe(ctx, request.(AuthMeRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "AuthMe")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(AuthMeResponseObject); ok {
		if err := validResponse.VisitAuthMeResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// AuthSignup operation middleware
func (sh *strictHandler) AuthSignup(w http.ResponseWriter, r *http.Request) {
	var request AuthSignupRequestObject

	var body AuthSignupJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
		return
	}
	request.Body = &body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.AuthSignup(ctx, request.(AuthSignupRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "AuthSignup")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(AuthSignupResponseObject); ok {
		if err := validResponse.VisitAuthSignupResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// ProjectsCreate operation middleware
func (sh *strictHandler) ProjectsCreate(w http.ResponseWriter, r *http.Request) {
	var request ProjectsCreateRequestObject

	var body ProjectsCreateJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
		return
	}
	request.Body = &body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.ProjectsCreate(ctx, request.(ProjectsCreateRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "ProjectsCreate")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(ProjectsCreateResponseObject); ok {
		if err := validResponse.VisitProjectsCreateResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}